requirements.txt
fake_useragent==1.4.0
MechanicalSoup==1.3.0
openpyxl==3.1.2
pandas==2.1.4
python-decouple==3.8
Requests==2.31.0
